+++ 2_fase/barcos fixos/uma_interação/plot_1.png
