+++ 2_fase/barcos_random/n_interções/plot_10.png
