+++ 2_fase/barcos_random/n_interções/plot_100.png
