+++ 2_fase/barcos_random/n_interções/plot_1000.png
